+++ compiler/rustc_query_system/src/dep_graph/mod.rs
@@ -8,8 +8,9 @@ mod serialized;
 use std::panic;
 
 pub use dep_node::{DepKind, DepKindVTable, DepNode, DepNodeKey, WorkProductId};
-pub(crate) use graph::DepGraphData;
-pub use graph::{DepGraph, DepNodeIndex, TaskDepsRef, WorkProduct, WorkProductMap, hash_result};
+pub use graph::{
+    DepGraph, DepGraphData, DepNodeIndex, TaskDepsRef, WorkProduct, WorkProductMap, hash_result,
+};
 pub use query::DepGraphQuery;
 use rustc_data_structures::profiling::SelfProfilerRef;
 use rustc_data_structures::sync::DynSync;
