+++ compiler/rustc_query_system/src/query/mod.rs
@@ -15,8 +15,8 @@ use rustc_span::def_id::DefId;
 pub use self::caches::{DefIdCache, DefaultCache, QueryCache, SingleCache, VecCache};
 pub use self::dispatcher::{HashResult, QueryDispatcher};
 pub use self::job::{
-    QueryInfo, QueryJob, QueryJobId, QueryJobInfo, QueryMap, break_query_cycles, print_query_stack,
-    report_cycle,
+    QueryInfo, QueryJob, QueryJobId, QueryJobInfo, QueryLatch, QueryMap, break_query_cycles,
+    print_query_stack, report_cycle,
 };
 pub use self::plumbing::*;
 use crate::dep_graph::{DepKind, DepNodeIndex, HasDepContext, SerializedDepNodeIndex};
@@ -84,7 +84,7 @@ impl<'tcx> QueryStackFrame<QueryStackDeferred<'tcx>> {
 #[derive(Clone, Debug)]
 pub struct QueryStackFrameExtra {
     pub description: String,
-    span: Option<Span>,
+    pub span: Option<Span>,
     pub def_kind: Option<DefKind>,
 }
